requirements.txt
# ── Backend API ──
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4

# ── Auth / Security ──
PyJWT==2.8.0
passlib==1.7.4

# ── Task Queue ──
celery==5.4.0
redis==5.0.7

# ── System Profiling ──
psutil==5.9.8
GPUtil==1.4.0

# ── Image / Video Processing ──
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# ── GPU / ML ──
torch>=2.2.0
torchvision>=0.17.0

# ── Distributed Computing ──
ray[default]==2.31.0

# ── Cloud SDK ──
boto3==1.34.131

# ── WebSocket ──
websockets==12.0

# ── Utilities ──
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==24.1.0
